pyproject.toml
[project]
name = "textual-agent-app"
version = "0.1.0"
description = "A TUI application using Textual with Google ADK integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=13.0.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "google-adk",
    "litellm>=1.74.8",
    "ollama>=0.5.1",
]

[project.optional-dependencies]
dev = ["textual[dev]>=0.41.0", "pytest>=7.0.0", "black>=23.0.0", "ruff>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py312"
